refactor: 데스크톱 앱 컴포넌트 명칭 정리
